Move mouse/keyboard models into their own files
