+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/KeyboardInputActionData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 639e76e528d04c60a85b31bed2c8674b
+timeCreated: 1718116068
