+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/MouseInputActionData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7f1586da674c43a6a400a003daaed00c
+timeCreated: 1718116125
